+++ GVRf/Framework/framework/src/main/java/org/gearvrf/io/GVRInputManager.java
@@ -24,11 +24,17 @@ import android.view.InputDevice;
 import android.view.KeyEvent;
 import android.view.MotionEvent;
 
+import org.gearvrf.GVRAndroidResource;
 import org.gearvrf.GVRContext;
 import org.gearvrf.GVRCursorController;
 import org.gearvrf.GVREventReceiver;
+import org.gearvrf.GVRRenderData;
+import org.gearvrf.GVRScene;
+import org.gearvrf.GVRSceneObject;
 import org.gearvrf.IEventReceiver;
 import org.gearvrf.IEvents;
+import org.gearvrf.R;
+import org.gearvrf.utility.Log;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -49,7 +55,7 @@ import java.util.List;
  * The class also allows external input devices to be added using the
  * {@link GVRInputManager#addCursorController(GVRCursorController)} method.
  */
-public abstract class GVRInputManager implements IEventReceiver
+public class GVRInputManager implements IEventReceiver
 {
     private static final String TAG = GVRInputManager.class.getSimpleName();
     private static final String WEAR_TOUCH_PAD_SERVICE_PACKAGE_NAME = "org.gearvrf.weartouchpad";
@@ -92,8 +98,12 @@ public abstract class GVRInputManager implements IEventReceiver
     // We make use of the vendor and product id to identify a device.
     private final SparseArray<GVRCursorController> cache;
 
-    protected GVRInputManager(GVRContext context, boolean useGazeCursorController,
-                              boolean useAndroidWearTouchpad) {
+    private List<CursorControllerListener> listeners;
+    private List<GVRCursorController> controllers;
+
+    public GVRInputManager(GVRContext context, boolean useGazeCursorController,
+                              boolean useAndroidWearTouchpad)
+    {
         Context androidContext = context.getContext();
         inputManager = (InputManager) androidContext
                 .getSystemService(Context.INPUT_SERVICE);
@@ -105,23 +115,196 @@ public abstract class GVRInputManager implements IEventReceiver
         cache = new SparseArray<GVRCursorController>();
         mouseDeviceManager = new GVRMouseDeviceManager(context);
         gamepadDeviceManager = new GVRGamepadDeviceManager();
+        gearCursorController = new GearCursorController(context);
         if(useAndroidWearTouchpad && checkIfWearTouchPadServiceInstalled(context)) {
             androidWearTouchpad = new GVRAndroidWearTouchpad(context);
         }
-        scanDevices();
+        controllers = new ArrayList<GVRCursorController>();
+        listeners = new ArrayList<CursorControllerListener>();
     }
 
     public GVREventReceiver getEventReceiver() { return mListeners; }
 
-    public abstract void scanControllers();
+    /**
+     * Emit "onCursorControllerAdded" events to controller listeners
+     * for all connected controllers.
+     * <p>
+     * To connect with controllers, the application must add a ControllerEventListener
+     * to listen for onCursorControllerAdded and onCursorControllerRemoved events
+     * and then call this function to emit the events.
+     * An event is sent immediately if the controller is actually ready
+     * and connected. If the controller connects later, the event
+     * will occur later and will not be sent from this function
+     * (this can happen with Bluetooth devices).
+     * @see CursorControllerListener
+     */
+    public void scanControllers()
+    {
+        for (GVRCursorController controller : getCursorControllers())
+        {
+            controllers.add(controller);
+            if (controller.isConnected())
+            {
+                addCursorController(controller);
+            }
+        }
+    }
 
-    public abstract void selectController(GVRContext ctx, GVRControllerType[] desiredTypes, ICursorControllerSelectListener listener);
+    /**
+     * Select an input controller based on a prioritized list of controller types.
+     * <p>
+     * The "onCursorControllerSelected" event is emitted when
+     * a cursor controller is chosen. The controller chosen is
+     * the highest priority controller available when the call is made.
+     * <p>
+     * If a higher priority controller is connected afterwards,
+     * the input manager switches to using the new controller
+     * and "onCursorControllerSelected" is emitted again.
+     * @param desiredTypes list of accepable controller types in prioritized
+     *                     order, highest priority is last in the list.
+     * @param listener     listens for onCursorControllerSelected events.
+     * @see CursorControllerListener
+     * @see ICursorControllerSelectListener
+     * @see org.gearvrf.io.GVRInputManager.ICursorControllerSelectListener
+     */
+    public void selectController(GVRControllerType[] desiredTypes, ICursorControllerSelectListener listener)
+    {
+        GVRInputManager.SingleControllerSelector
+                selector = new GVRInputManager.SingleControllerSelector(context, desiredTypes);
+        addCursorControllerListener(selector);
+        getEventReceiver().addListener(listener);
+        scanControllers();
+    }
 
-    private void scanDevices()
+    /**
+     * Select an input controller based on a prioritized list type names.
+     * <p>
+     * The "onCursorControllerSelected" event is emitted when
+     * a cursor controller is chosen. The controller chosen is
+     * the highest priority controller available when the call is made.
+     * <p>
+     * If a higher priority controller is connected afterwards,
+     * the input manager switches to using the new controller
+     * and "onCursorControllerSelected" is emitted again.
+     * <p>
+     * This form of the function is useful when using JavaScript.
+     * The event is routed to the event receiver of the {@link GVRInputManager}
+     * and can be handled by attaching a script which contains a
+     * function called "onCursorControllerSelected".
+     * </p>
+     * @param typeNames list of accepable controller type names in prioritized
+     *                     order, highest priority is last in the list.
+     * @see ICursorControllerSelectListener
+     * @see org.gearvrf.io.GVRInputManager.ICursorControllerSelectListener
+     */
+    public void selectController(String[] typeNames)
+    {
+        GVRControllerType[] desiredTypes = new GVRControllerType[typeNames.length];
+
+        for (String name : typeNames)
+        {
+            String typeName = name.toUpperCase();
+            int i = 0;
+            for (GVRControllerType type : GVRControllerType.values())
+            {
+                if (type.name().equals(typeName))
+                {
+                    desiredTypes[i] = type;
+                    break;
+                }
+            }
+        }
+        GVRInputManager.SingleControllerSelector
+                selector = new GVRInputManager.SingleControllerSelector(context, desiredTypes);
+        addCursorControllerListener(selector);
+        scanControllers();
+    }
+
+    /**
+     * Add a {@link CursorControllerListener} to receive an event whenever a
+     * {@link GVRCursorController} is added or removed from the framework.
+     *
+     * @param listener the {@link CursorControllerListener} to be added.
+     */
+    public void addCursorControllerListener(CursorControllerListener listener) {
+        synchronized (listeners) {
+            listeners.add(listener);
+        }
+    }
+
+    /**
+     * Remove the previously added {@link CursorControllerListener}.
+     *
+     * @param listener the {@link CursorControllerListener} to be removed.
+     */
+    public void removeCursorControllerListener(
+            CursorControllerListener listener) {
+        synchronized (listeners) {
+            listeners.remove(listener);
+        }
+    }
+
+    /**
+     * Define a {@link GVRCursorController} and add it to the
+     * {@link GVRInputManager} for external input device handling by the
+     * framework.
+     *
+     * @param controller the external {@link GVRCursorController} to be added to the
+     *                   framework.
+     */
+    public void addCursorController(GVRCursorController controller) {
+        synchronized (listeners) {
+            for (CursorControllerListener listener : listeners) {
+                listener.onCursorControllerAdded(controller);
+            }
+        }
+    }
+
+
+    /**
+     * Remove the previously added {@link GVRCursorController} added to the
+     * framework.
+     *
+     * @param controller the external {@link GVRCursorController} to be removed from
+     *                   the framework.
+     */
+    public void removeCursorController(GVRCursorController controller) {
+        controllers.remove(controller);
+        synchronized (listeners) {
+            for (CursorControllerListener listener : listeners) {
+                listener.onCursorControllerRemoved(controller);
+            }
+        }
+    }
+
+
+    /**
+     * This method sets a new scene for the {@link GVRInputManager}
+     *
+     * @param scene
+     *
+     */
+    public void setScene(GVRScene scene) {
+        for (GVRCursorController controller : controllers) {
+            controller.setScene(scene);
+            controller.invalidate();
+        }
+    }
+
+    /**
+     * Scan all the attached Android IO devices and gather
+     * information about them.
+     *
+     * This function is called once at initialization to enumerate
+     * the Android devices. Calling it again is not harmful but
+     * will not gather additional information.
+     */
+    public void scanDevices()
     {
         for (int deviceId : inputManager.getInputDeviceIds()) {
             addDevice(deviceId);
         }
+        cache.put(CONTROLLER_CACHED_KEY, gearCursorController);
     }
 
     private boolean checkIfWearTouchPadServiceInstalled(GVRContext context) {
@@ -191,7 +374,13 @@ public abstract class GVRInputManager implements IEventReceiver
         return false;
     }
 
-    protected void close() {
+    /**
+     * Shut down the input manager.
+     *
+     * After this call, GearVRf will not be able to access IO devices.
+     */
+    public void close()
+    {
         inputManager.unregisterInputDeviceListener(inputDeviceListener);
         controllerIds.clear();
         cache.clear();
@@ -200,6 +389,7 @@ public abstract class GVRInputManager implements IEventReceiver
         if (gazeCursorController != null) {
             gazeCursorController.close();
         }
+        controllers.clear();
     }
 
     // returns null if no device is found.
@@ -223,17 +413,10 @@ public abstract class GVRInputManager implements IEventReceiver
 
         int vendorId = device.getVendorId();
         int productId = device.getProductId();
-        boolean isTouchScreen = ((sources & InputDevice.SOURCE_TOUCHSCREEN) == InputDevice.SOURCE_TOUCHSCREEN);
         boolean isKeyBoard = ((sources & InputDevice.SOURCE_KEYBOARD) == InputDevice.SOURCE_KEYBOARD);
         boolean isTouchPad =   ((sources & InputDevice.SOURCE_TOUCHPAD) == InputDevice.SOURCE_TOUCHPAD);
         boolean isMouse =   ((sources & InputDevice.SOURCE_MOUSE) == InputDevice.SOURCE_MOUSE);
 
-        if (sources == (InputDevice.SOURCE_TOUCHSCREEN | InputDevice.SOURCE_KEYBOARD)) {
-            return GVRControllerType.CONTROLLER;
-        }
-        if (isTouchScreen && isKeyBoard) {
-            return GVRControllerType.CONTROLLER;
-        }
         if (isKeyBoard || isTouchPad) {
             // Allow gpio keyboard to be a gaze controller if enabled, also allow
             // any keyboard/touchpad device without a product/vendor id (assumed to be
@@ -254,7 +437,6 @@ public abstract class GVRInputManager implements IEventReceiver
             }
             return GVRControllerType.MOUSE;
         }
-
         return GVRControllerType.UNKNOWN;
     }
 
@@ -285,10 +467,7 @@ public abstract class GVRInputManager implements IEventReceiver
         }
 
         int key;
-        if (controllerType == GVRControllerType.CONTROLLER) {
-            key =  CONTROLLER_CACHED_KEY;
-        }
-        else if (controllerType == GVRControllerType.GAZE) {
+        if (controllerType == GVRControllerType.GAZE) {
             // create the controller if there isn't one. 
             if (gazeCursorController == null) {
                 gazeCursorController = new GVRGazeCursorController(context, GVRControllerType.GAZE,
@@ -303,27 +482,29 @@ public abstract class GVRInputManager implements IEventReceiver
             key = getCacheKey(device, controllerType);
         }
 
-        if (key != -1) {
+        if (key != -1)
+        {
             GVRCursorController controller = cache.get(key);
-            if (controller == null) {
-                if (controllerType == GVRControllerType.CONTROLLER) {
-                    controller = gearCursorController = new GearCursorController(context);
+            if (controller == null)
+            {
+                if (controllerType == GVRControllerType.MOUSE)
+                {
+                    controller = mouseDeviceManager.getCursorController(context, device.getName(), device.getVendorId(), device.getProductId());
                 }
-                if (controllerType == GVRControllerType.MOUSE) {
-                    controller = mouseDeviceManager
-                            .getCursorController(context, device.getName(), device.getVendorId(),
-                                    device.getProductId());
-                } else if (controllerType == GVRControllerType.GAMEPAD) {
-                    controller = gamepadDeviceManager
-                            .getCursorController(context, device.getName(), device.getVendorId(),
-                                    device.getProductId());
-                } else if (controllerType == GVRControllerType.GAZE) {
+                else if (controllerType == GVRControllerType.GAMEPAD)
+                {
+                    controller = gamepadDeviceManager.getCursorController(context, device.getName(), device.getVendorId(), device.getProductId());
+                }
+                else if (controllerType == GVRControllerType.GAZE)
+                {
                     controller = gazeCursorController;
                 }
                 cache.put(key, controller);
                 controllerIds.put(device.getId(), controller);
                 return controller;
-            } else {
+            }
+            else
+            {
                 controllerIds.put(device.getId(), controller);
             }
         }
@@ -366,33 +547,6 @@ public abstract class GVRInputManager implements IEventReceiver
         return null;
     }
 
-    private InputDeviceListener inputDeviceListener = new InputDeviceListener() {
-
-        @Override
-        public void onInputDeviceRemoved(int deviceId) {
-            GVRCursorController controller = removeDevice(deviceId);
-            if (controller != null) {
-                controller.setScene(null);
-                controller.setEnable(false);
-            }
-        }
-
-        @Override
-        public void onInputDeviceChanged(int deviceId) {
-            // TODO: Not Used, see if needed.
-        }
-
-        @Override
-        public void onInputDeviceAdded(int deviceId) {
-            GVRCursorController controller = addDevice(deviceId);
-            if (controller != null) {
-                controller.setScene(context.getMainScene());
-                controller.setEnable(true);
-                addCursorController(controller);
-            }
-        }
-    };
-
     /**
      * Dispatch a {@link KeyEvent} to the {@link GVRInputManager}.
      *
@@ -423,39 +577,32 @@ public abstract class GVRInputManager implements IEventReceiver
         return false;
     }
 
-    /**
-     * Add a {@link CursorControllerListener} to receive an event whenever a
-     * {@link GVRCursorController} is added or removed from the framework.
-     *
-     * @param listener the {@link CursorControllerListener} to be added.
-     */
-    public abstract void addCursorControllerListener(CursorControllerListener listener);
+    private InputDeviceListener inputDeviceListener = new InputDeviceListener() {
 
-    /**
-     * Remove the previously added {@link CursorControllerListener}.
-     *
-     * @param listener the {@link CursorControllerListener} to be removed.
-     */
-    public abstract void removeCursorControllerListener(CursorControllerListener listener);
+        @Override
+        public void onInputDeviceRemoved(int deviceId) {
+            GVRCursorController controller = removeDevice(deviceId);
+            if (controller != null) {
+                controller.setEnable(false);
+            }
+        }
 
-    /**
-     * Define a {@link GVRCursorController} and add it to the
-     * {@link GVRInputManager} for external input device handling by the
-     * framework.
-     *
-     * @param controller the external {@link GVRCursorController} to be added to the
-     *                   framework.
-     */
-    public abstract void addCursorController(GVRCursorController controller);
+        @Override
+        public void onInputDeviceChanged(int deviceId) {
+            // TODO: Not Used, see if needed.
+        }
 
-    /**
-     * Remove the previously added {@link GVRCursorController} added to the
-     * framework.
-     *
-     * @param controller the external {@link GVRCursorController} to be removed from
-     *                   the framework.
-     */
-    public abstract void removeCursorController(GVRCursorController controller);
+        @Override
+        public void onInputDeviceAdded(int deviceId) {
+            GVRCursorController controller = addDevice(deviceId);
+            if (controller != null) {
+                controller.setScene(context.getMainScene());
+                controller.setEnable(true);
+                addCursorController(controller);
+            }
+        }
+    };
+
 
     /**
      * Interface to listen for cursor controller selection events.
@@ -465,4 +612,132 @@ public abstract class GVRInputManager implements IEventReceiver
     {
         public void onCursorControllerSelected(GVRCursorController newController, GVRCursorController oldController);
     }
+
+    protected static class SingleControllerSelector implements CursorControllerListener
+    {
+        private GVRControllerType[] mControllerTypes;
+        private int mCurrentControllerPriority = -1;
+        private GVRCursorController mCursorController = null;
+        private GVRSceneObject mCursor = null;
+        private GVRCursorController.CursorControl mCursorControl = GVRCursorController.CursorControl.CURSOR_CONSTANT_DEPTH;
+
+        public SingleControllerSelector(GVRContext ctx, GVRControllerType[] desiredTypes)
+        {
+            mControllerTypes = desiredTypes;
+            mCursor = makeDefaultCursor(ctx);
+        }
+
+        @Override
+        synchronized public void onCursorControllerAdded(GVRCursorController gvrCursorController)
+        {
+            if (mCursorController == gvrCursorController)
+            {
+                return;
+            }
+
+            int priority = getControllerPriority(gvrCursorController.getControllerType());
+            if (priority > mCurrentControllerPriority)
+            {
+                deselectController();
+                selectController(gvrCursorController);
+                if (gvrCursorController instanceof GearCursorController)
+                {
+                    ((GearCursorController) gvrCursorController).showControllerModel(true);
+                }
+                mCurrentControllerPriority = priority;
+            }
+            else
+            {
+                gvrCursorController.setEnable(false);
+            }
+        }
+
+        public GVRCursorController getController()
+        {
+            return mCursorController;
+        }
+
+        public GVRSceneObject getCursor() { return mCursor; }
+
+        public void setCursor(GVRSceneObject cursor)
+        {
+            mCursor = cursor;
+        }
+
+        private GVRSceneObject makeDefaultCursor(GVRContext ctx)
+        {
+            GVRSceneObject cursor = new GVRSceneObject(ctx, 1, 1,
+                                                       ctx.getAssetLoader().loadTexture(
+                                                               new GVRAndroidResource(ctx, R.drawable.cursor)));
+            GVRRenderData rdata = cursor.getRenderData();
+            rdata.setDepthTest(false);
+            rdata.disableLight();
+            rdata.setRenderingOrder(GVRRenderData.GVRRenderingOrder.OVERLAY + 10);
+            cursor.getTransform().setScale(0.2f, 0.2f, 1.0f);
+            return cursor;
+        }
+
+        private void selectController(GVRCursorController controller)
+        {
+            GVRContext ctx = controller.getGVRContext();
+            controller.setScene(controller.getGVRContext().getMainScene());
+            if (mCursor != null)
+            {
+                mCursor.getTransform().setPosition(0, 0, 0);
+                controller.setCursor(mCursor);
+            }
+            ctx.getEventManager().sendEvent(ctx.getInputManager(),
+                                            ICursorControllerSelectListener.class,
+                                            "onCursorControllerSelected",
+                                            controller,
+                                            mCursorController);
+            mCursorController = controller;
+            mCursorController.setEnable(true);
+            Log.d(TAG, "selected " + controller.getClass().getSimpleName());
+        }
+
+        private void deselectController()
+        {
+            GVRCursorController c = mCursorController;
+            if (c != null)
+            {
+                mCursorController = null;
+                mCurrentControllerPriority = -1;
+                c.setCursor(null);
+                c.setEnable(false);
+            }
+        }
+
+        public void onCursorControllerRemoved(GVRCursorController gvrCursorController)
+        {
+            /*
+             * If we are removing the controller currently being used,
+             * switch to the Gaze controller if possible
+             */
+            if (mCursorController == gvrCursorController)
+            {
+                GVRContext ctx = gvrCursorController.getGVRContext();
+                deselectController();
+                GVRCursorController gaze = ctx.getInputManager().findCursorController(GVRControllerType.GAZE);
+                if (gaze != null)
+                {
+                    ctx.getInputManager().addCursorController(gaze);
+                }
+            }
+        }
+
+        private int getControllerPriority(GVRControllerType type)
+        {
+            int i = 0;
+            for (GVRControllerType t : mControllerTypes)
+            {
+                if (t.equals(type))
+                {
+                    return i;
+                }
+                ++i;
+            }
+            return -1;
+        }
+    };
 }
